pyproject.toml
[project]
name = "link-persona-bot"
version = "0.1.0"
description = "Discord Bot for URL-shared article summarization with AI personalities"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "discord.py>=2.4.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "trafilatura>=1.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "bot"]

[tool.uv]
dev-dependencies = []
